### 2018-01-30 #559 @whedon
[PRE REVIEW]: ChebTools: C++11 (and Python) tools for working with Chebyshev expansions
**Submitting author:** @ianhbell (<a href="http://orcid.org/0000-0003-1091-9080">Ian Bell</a>)
**Repository:** <a href="https://github.com/usnistgov/ChebTools" target ="_blank">https://github.com/usnistgov/ChebTools</a>
**Version:** v1.0.0
