### 2018-01-30 #559 @arfon
@ianhbell - thanks for your submission to JOSS. From a quick inspection of the source code on GitHub it looks like this software is released under some kind of public domain notice.

Unfortunately we don't accept submissions that don't have an [OSI-approved open source license](https://opensource.org/licenses).

I realize this is potentially an issue for government employees and we've discussed this at length already here. Please take a look at this issue for background and let me know if you think you might be able to add a license to this project: https://github.com/openjournals/joss/issues/179
