### 2018-01-30 #559 @ianhbell
I just spoke with legal at NIST, and it seems that we can do a dual-license: public domain (whatever that *actually* means) in the USA, and something else (MIT?) elsewhere.  At least that was the most recent ruling I got.  I'll be in touch just as soon as I have some new license verbiage.
