### 2018-01-30 #559 @ianhbell
Update: License updated to MIT-ish.

As this is my first (of hopefully many) submissions to JOSS, I didn't fully understand the two-sentence snippet in the submission would be visible to the public on the paper submission record.  Any chance you could remove the P.S. that I added to my submission?  I don't seem to be able to.
