### 2018-01-30 #559 @ianhbell
Some recommended reviewers: @drvinceknight @jordigh @JaroslavHron
