### 2018-01-30 #559 @drvinceknight
I would be happy to review. To the editor: my knowledge of C++ is not great but I'm familiar with Python and the field.
