### 2018-02-04 #559 @labarba
I can edit this one, @arfon -- one question: the "modified MIT" license ... is it really _modified_ from MIT? Because if so, it would not be OSI-approved. I has to be the _actual_ MIT license, no?
