### 2018-02-06 #559 @ianhbell
Pinged our legal folks, a clean MIT license plus a note in the README seems sufficient, but I need final confirmation.
