### 2018-02-06 #559 @arfon
> Pinged our legal folks, a clean MIT license plus a note in the README seems sufficient, but I need final confirmation.

Excellent. Thanks @ianhbell
