### 2018-02-06 f8ef34d
Modify license and README one more time
