### 2018-02-06 #559 @arfon
> One more try... Does this work?

If you could just remove the 'Terms of use' section at the top of the `LICENSE` file I think we'll be set: https://github.com/usnistgov/ChebTools/blob/master/LICENSE#L1-L4
