### 2018-02-06 #559 @arfon
> I can edit this one, @arfon -- one question: the "modified MIT" license ... is it really modified from MIT? Because if so, it would not be OSI-approved. I has to be the actual MIT license, no?

@labarba - thanks for volunteering to edit this submission. I think we've resolved the licensing concerns here now.
