### 2018-02-06 #559 @ianhbell
It's no longer a modified MIT license, though it was before.  Working for the US Government and writing open-source software is at times a frustrating exercise...
