### 2018-02-07 #569 @labarba
OK, @drvinceknight: here's where the action happens. If you've reviewed for JOSS before, you know the drill. If not, I'm here to answer your questions!
