### 2018-02-07 #569 @labarba
@ianhbell -- Your paper is not compiling. Can you check on that?
