### 2018-02-07 #569 @ianhbell
I don't think that is related to my paper.  I don't do anything too crazy, perhaps it is related to my Python code block?  I don't have any way of testing compilation on my side.  I asked several times about that but was never able to get a working build to the journal specs on my machine.  I think the whedon server needs to have the LaTeX package added.
