### 2018-02-08 #569 @arfon
@labarba @ianhbell - looks like our @whedon LaTeX environment doesn't include `fancyvrb` (which is probably associated with the code block). I've added it in https://github.com/openjournals/heroku-buildpack-tex/commit/5563891f668a8b7dd62fd67f740a79a5e80772d4 but this won't get deployed for a while.

For now will this copy of the paper work that I compiled locally? [10.21105.joss.00569.pdf](https://github.com/openjournals/joss-reviews/files/1706802/10.21105.joss.00569.pdf)
