### 2018-02-08 #569 @ianhbell
That local copy looks beautiful!

How about making a docker container with whedon inside?  That might be the cleanest packaging.  I can try to help with that if you are interested?
