### 2018-02-08 #569 @drvinceknight
> OK, @drvinceknight: here's where the action happens. If you've reviewed for JOSS before, you know the drill. If not, I'm here to answer your questions!

I haven't reviewed for JOSS before but everything looks pretty self explanatory (I'm just sat here thinking how great the whole system is!).

I'm aiming to complete this over next couple of days.
