### 2018-02-09 #569 @drvinceknight
## Installation

1. `python setup.py install` fails if `cmake` is not installed (on linux), could we add that to the `README.md` (`apt install cmake` solves this without a problem).
2. Once `cmake` is installed `python setup.py install` still fails, here's the tail of the output:

   ````
   -- Performing Test OpenMP_FLAG_DETECTED - Success
   -- Found OpenMP: -fopenmp
   CMake Error at CMakeLists.txt:32 (add_subdirectory):
     The source directory

       /home/vince/Desktop/ChebTools/externals/pybind11

     does not contain a CMakeLists.txt file.

   CMake Error at CMakeLists.txt:33 (pybind11_add_module):
     Unknown CMake command "pybind11_add_module".

   -- Configuring incomplete, errors occurred!
   See also "/home/vince/Desktop/ChebTools/build/temp.linux-x86_64-3.5/CMakeFiles/CMakeOutput.log".
   Traceback (most recent call last):
     File "setup.py", line 70, in <module>
       zip_safe=False,
     File "/home/vince/anaconda3/lib/python3.5/site-packages/setuptools/__init__.py", line 129, in setup
       return distutils.core.setup(**attrs)
     File "/home/vince/anaconda3/lib/python3.5/distutils/core.py", line 148, in setup
       dist.run_commands()
     File "/home/vince/anaconda3/lib/python3.5/distutils/dist.py", line 955, in run_commands
       self.run_command(cmd)
     File "/home/vince/anaconda3/lib/python3.5/distutils/dist.py", line 974, in run_command
       cmd_obj.run()
     File "/home/vince/anaconda3/lib/python3.5/site-packages/setuptools/command/develop.py", line 36, in run
       self.install_for_development()
     File "/home/vince/anaconda3/lib/python3.5/site-packages/setuptools/command/develop.py", line 136, in install_for_development
       self.run_command('build_ext')
     File "/home/vince/anaconda3/lib/python3.5/distutils/cmd.py", line 313, in run_command
       self.distribution.run_command(command)
     File "/home/vince/anaconda3/lib/python3.5/distutils/dist.py", line 974, in run_command
       cmd_obj.run()
     File "setup.py", line 34, in run
       self.build_extension(ext)
     File "setup.py", line 58, in build_extension
       subprocess.check_call(['cmake', ext.sourcedir] + cmake_args, cwd=self.build_temp, env=env)
     File "/home/vince/anaconda3/lib/python3.5/subprocess.py", line 581, in check_call
       raise CalledProcessError(retcode, cmd)
   subprocess.CalledProcessError: Command '['cmake', '/home/vince/Desktop/ChebTools', '   -DCMAKE_LIBRARY_OUTPUT_DIRECTORY=/home/vince/Desktop/ChebTools', '-
   DPYTHON_EXECUTABLE=/home/vince/anaconda3/bin/python', '-DCMAKE_BUILD_TYPE=Release']'
   returned non-zero exit status 1
   ```
3. All the instructions in https://github.com/usnistgov/ChebTools#cmake-build work fine and following those running `python setup.py install` reports no error however there is still a problem when importing the library:

   ```
   Desktop: python
   Python 3.5.2 |Anaconda custom (64-bit)| (default, Jul  2 2016, 17:53:06)
   [GCC 4.4.7 20120313 (Red Hat 4.4.7-1)] on linux
    >>> import ChebTools
   Traceback (most recent call last):
     File "<stdin>", line 1, in <module>
   ImportError: /home/vince/anaconda3/bin/../lib/libstdc++.so.6: version `GLIBCXX_3.4.20' not found
   (required by /home/vince/Desktop/ChebTools/ChebTools.cpython-35m-x86_64-linux-gnu.so)
   >>>
   ```

My suggestion would be to modify the install instructions to make the `cmake-build` instructions the standard way, currently I think "Install in one line from github" https://github.com/usnistgov/ChebTools#to-install-in-one-line-from-github is a bit misleading?

## Documentation

As far as I can see there's no simple example usage in the `README.md`. I know there is a binder link but also having a minimum example in the `README.md` is necessary. The small example that's in the paper would be fine.

There is also no "A statement of need" in the documentation (`README.md`), the statement of need in the paper is great.

## Community

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This seems to be missing.

## Release version

@labarba I'm not entirely sure I follow this:

> Version: Does the release version given match the GitHub release (v1.0.0)?

As far as I can tell this software is "only" released on github. Is this requiring/expecting the software itself to be archived on a service like Zenodo?

---

I haven't been able to test any of the Functionality aspects, will do that once I've understood what I've missed with the installation. @ianhbell I hope this is helpful and please let me know if I've misunderstood anything.
