### 2018-02-09 #569 @labarba
@drvinceknight : The software will need to be archived on Zenodo or similar, and get a DOI. But we've relaxed the requirement to have that _before_ acceptance, because when the review elicits changes, this leads to a new release. We ask that authors deposit the accepted version.
