### 2018-02-09 c3cad52
Update the docs according to JOSS review
