### 2018-02-09 6d4d7f6
Clean up formatting a bit?
