### 2018-02-09 #569 @ianhbell
@drvinceknight - the build problem you had was almost certainly because you didn't do a recursive git clone - I've been bitten by that innumerable times.  I updated the docs to try to address that issue and I tried to address your other concerns as well.

What do you think now: https://github.com/usnistgov/ChebTools ?

Ready for any other review you care to offer.
