### 2018-02-09 #569 @ianhbell
And yes, I have minted a doi on zenodo already, but I'll remake the release when the review is complete.
