### 2018-02-09 #569 @drvinceknight
@ianhbell the docs look great! Thanks for those changes.

I'm still having issues installing on my system though:

```
Desktop: pip install git+git://github.com/usnistgov/ChebTools.git
Collecting git+git://github.com/usnistgov/ChebTools.git
  Cloning git://github.com/usnistgov/ChebTools.git to /tmp/pip-yimzocuc-build
Installing collected packages: ChebTools
  Running setup.py install for ChebTools ... done
Successfully installed ChebTools-1.0.0
You are using pip version 8.1.2, however version 9.0.1 is available.
You should consider upgrading via the 'pip install --upgrade pip' command.
Desktop: python
Python 3.5.2 |Anaconda custom (64-bit)| (default, Jul  2 2016, 17:53:06)
[GCC 4.4.7 20120313 (Red Hat 4.4.7-1)] on linux
>>> import ChebTool
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: No module named 'ChebTool'
>>> import ChebTools
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: /home/vince/anaconda3/bin/../lib/libstdc++.so.6: version `GLIBCXX_3.4.20' not found (required by /home/vince/anaconda3/lib/python3.5/site-packages/ChebTools.cpython-35m-x86_64-linux-gnu.so)

```

I went for the `pip install git+git://github.com/usnistgov/ChebTools.git` approach which succeeded but as you can see I still seem to have some issues with it finding some file.
