### 2018-02-09 #569 @ianhbell
Sorry I forgot to answer about that import error.  That's a bug in anaconda... : https://github.com/ContinuumIO/anaconda-issues/issues/483

The short answer is : delete the libstdc++ that it references, and it will then hook your system GLIBC and it should all work fine, though you *might* want to backup your libstdc++ before you do that :)
