### 2018-02-09 #569 @drvinceknight
Running `conda install libgcc` fixed it for me (is it worth including something about that in the README perhaps?).
