### 2018-02-09 #569 @drvinceknight
Thanks for the changes @ianhbell! Nice piece of software and a nice paper :)

I have now ticked all boxes @labarba (except the one about the version but that all seems fine to me). Is there anything more formal I should do at this point? :)
