### 2018-02-09 502082a
Added note about problem with libstdc++
