### 2018-02-12 #569 @drvinceknight
> I take it your recommendation is to accept the submission, at this point, @drvinceknight ?

Yes, my recommendation is to accept the submission.
