### 2018-02-12 #569 @labarba
@ianhbell : Can you update the release, up the DOI version in Zenodo, and post the DOI right here, so we can link it to the article?

@arfon : This one is ready to publish.
